refactor: Remove commented out code for locking in RmFileHandle
This commit removes the commented out code for locking in the `RmFileHandle` class. The code was originally used to lock a record, but it has been disabled by commenting it out. Since the code is no longer needed, it is safe to remove it completely. This cleanup improves code readability and reduces unnecessary clutter.

Refactor code to remove commented out locking code in `RmFileHandle`.